separate logic to WeightedRandomSelector
