+++ src/test/java/net/datafaker/providers/base/CustomFakerTest.java
@@ -37,15 +37,15 @@ class CustomFakerTest {
             return INSECT_NAMES[faker.random().nextInt(INSECT_NAMES.length)];
         }
 
-        public String weightedInsectName() {
-            List<Map<String, Object>> insects = List.of(
-                Map.of("value", "Driver ant", "weight", 6.0),
-                Map.of("value", "Fire ant", "weight", 3.0),
-                Map.of("value", "Harvester ant", "weight", 1.0)
-            );
-
-            return faker.random().weightedArrayElement(insects);
-        }
+//        public String weightedInsectName() {
+//            List<Map<String, Object>> insects = List.of(
+//                Map.of("value", "Driver ant", "weight", 6.0),
+//                Map.of("value", "Fire ant", "weight", 3.0),
+//                Map.of("value", "Harvester ant", "weight", 1.0)
+//            );
+//
+//            return faker.random().weightedArrayElement(insects);
+//        }
     }
 
     public static class InsectFromFile extends AbstractProvider<BaseProviders> {
@@ -135,23 +135,23 @@ class CustomFakerTest {
         assertThat(insect1).isNotSameAs(insect2);
     }
 
-    @Test
-    void weightedInsectNameTest() {
-        MyCustomFaker myFaker = new MyCustomFaker();
-        Map<String, Integer> insectCounts = new HashMap<>();
-        insectCounts.put("Driver ant", 0);
-        insectCounts.put("Fire ant", 0);
-        insectCounts.put("Harvester ant", 0);
-
-        // Perform the weighted selection 100 times to check the distribution
-        for (int i = 0; i < 100; i++) {
-            String selectedInsect = myFaker.insect().weightedInsectName();
-            insectCounts.put(selectedInsect, insectCounts.get(selectedInsect) + 1);
-        }
-
-        assertThat(insectCounts.get("Driver ant")).isGreaterThan(insectCounts.get("Fire ant"));
-        assertThat(insectCounts.get("Fire ant")).isGreaterThan(insectCounts.get("Harvester ant"));
-
-        // insectCounts.forEach((insect, count) -> System.out.println(insect + ": " + count));
-    }
+//    @Test
+//    void weightedInsectNameTest() {
+//        MyCustomFaker myFaker = new MyCustomFaker();
+//        Map<String, Integer> insectCounts = new HashMap<>();
+//        insectCounts.put("Driver ant", 0);
+//        insectCounts.put("Fire ant", 0);
+//        insectCounts.put("Harvester ant", 0);
+//
+//        // Perform the weighted selection 100 times to check the distribution
+//        for (int i = 0; i < 100; i++) {
+//            String selectedInsect = myFaker.insect().weightedInsectName();
+//            insectCounts.put(selectedInsect, insectCounts.get(selectedInsect) + 1);
+//        }
+//
+//        assertThat(insectCounts.get("Driver ant")).isGreaterThan(insectCounts.get("Fire ant"));
+//        assertThat(insectCounts.get("Fire ant")).isGreaterThan(insectCounts.get("Harvester ant"));
+//
+//        // insectCounts.forEach((insect, count) -> System.out.println(insect + ": " + count));
+//    }
 }
