+++ src/test/java/net/datafaker/service/RandomServiceTest.java
@@ -1,7 +1,6 @@
 package net.datafaker.service;
 
 import net.datafaker.AbstractFakerTest;
-
 import org.assertj.core.api.Condition;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
@@ -9,11 +8,6 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Random;
 import java.util.stream.Stream;
 
@@ -25,225 +19,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
  * @author pmiklos
  */
 class RandomServiceTest extends AbstractFakerTest {
-
-
-    public static final String WEIGHT_KEY = "weight";
-    public static final String VALUE_KEY = "value";
-    public static final String ELEMENT_1 = "Element1";
-    public static final String ELEMENT_2 = "Element2";
-    public static final String ELEMENT_3 = "Element3";
-    public static final String STRING_WEIGHT = "1.0";
-    public static final int ITERATIONS = 1000;
-    public static final double ELEMENT_1_WEIGHT = 1.0;
-    public static final double ELEMENT_2_WEIGHT = 2.0;
-    public static final double ELEMENT_3_WEIGHT = 3.0;
-    public static final double ZERO_WEIGHT = 0.0;
-    public static final double NEGATIVE_WEIGHT = -1.0;
-
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withSingleElement(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, ELEMENT_1_WEIGHT)
-        );
-
-        String result = randomService.weightedArrayElement(items);
-        assertThat(result).isEqualTo(ELEMENT_1);
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withMultipleElements(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, ELEMENT_1_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_2_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_3, WEIGHT_KEY, ELEMENT_3_WEIGHT)
-        );
-
-        Map<String, Integer> counts = countResults(randomService, items);
-
-        assertThat(counts).containsKeys(ELEMENT_1, ELEMENT_2, ELEMENT_3);
-        assertThat(counts.get(ELEMENT_1)).isLessThan(counts.get(ELEMENT_2));
-        assertThat(counts.get(ELEMENT_2)).isLessThan(counts.get(ELEMENT_3));
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withNullItems(RandomService randomService) {
-        assertThatThrownBy(() -> randomService.weightedArrayElement(null))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("weightedArrayElement expects a non-empty list");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withEmptyItems(RandomService randomService) {
-        List<Map<String, Object>> items = Collections.emptyList();
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("weightedArrayElement expects a non-empty list");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withMissingWeightKey(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_2_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Each item must contain 'weight' and 'value' keys");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withMissingValueKey(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(WEIGHT_KEY, ELEMENT_1_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_2_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Each item must contain 'weight' and 'value' keys");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withInvalidWeightType(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, STRING_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_2_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Weight must be a non-null Double");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withNaNWeight(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, Double.NaN),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_2_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Weight must be a positive number and cannot be NaN or infinite");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withZeroWeight(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, ZERO_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_2_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Weight must be a positive number and cannot be NaN or infinite");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withNegativeWeight(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, NEGATIVE_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_2_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Weight must be a positive number and cannot be NaN or infinite");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withInfiniteWeight(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, Double.POSITIVE_INFINITY),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_2_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Weight must be a positive number and cannot be NaN or infinite");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withNullItem(RandomService randomService) {
-        List<Map<String, Object>> items = new ArrayList<>();
-        items.add(null);
-        items.add(Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, ELEMENT_1_WEIGHT));
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Item cannot be null or empty");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withEmptyItem(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(),
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, ELEMENT_1_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Item cannot be null or empty");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testWeightedArrayElement_withNullValue(RandomService randomService) {
-        List<Map<String, Object>> items = new ArrayList<>();
-
-        Map<String, Object> itemWithNullValue = new HashMap<>();
-        itemWithNullValue.put(VALUE_KEY, null);
-        itemWithNullValue.put(WEIGHT_KEY, ELEMENT_1_WEIGHT);
-
-        items.add(itemWithNullValue);
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Value cannot be null");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testCalculateTotalWeight_withOverflowingWeights(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, Double.MAX_VALUE - 1),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, Double.MAX_VALUE - 1)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Sum of the weights exceeds Double.MAX_VALUE");
-    }
-
-    @ParameterizedTest
-    @MethodSource("randomServiceProvider")
-    void testCalculateTotalWeight_withDuplicateKeys(RandomService randomService) {
-        List<Map<String, Object>> items = List.of(
-            Map.of(VALUE_KEY, ELEMENT_1, WEIGHT_KEY, ELEMENT_1_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_1_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_2, WEIGHT_KEY, ELEMENT_1_WEIGHT),
-            Map.of(VALUE_KEY, ELEMENT_3, WEIGHT_KEY, ELEMENT_1_WEIGHT)
-        );
-
-        assertThatThrownBy(() -> randomService.weightedArrayElement(items))
-            .isInstanceOf(IllegalArgumentException.class)
-            .hasMessage("Duplicate value found: Element2. Values must be unique.");
-    }
 
     @ParameterizedTest
     @MethodSource("randomServiceProvider")
@@ -368,15 +143,6 @@ class RandomServiceTest extends AbstractFakerTest {
     @MethodSource("randomServiceProvider")
     void testDefaultHex(RandomService randomService) {
         assertThat(randomService.hex()).matches("^[0-9A-F]{8}$");
-    }
-
-    private Map<String, Integer> countResults(RandomService randomService, List<Map<String, Object>> items) {
-        Map<String, Integer> counts = new HashMap<>();
-        for (int i = 0; i < ITERATIONS; i++) {
-            String result = randomService.weightedArrayElement(items);
-            counts.merge(result, 1, Integer::sum);
-        }
-        return counts;
     }
 
     private static Stream<Arguments> randomServiceProvider() {
